finish training with 110 epochs over 5 runs
Starting with the second run, number of GPUs was doubled to 8 (more wasn't possible due to resource constraints on the SLURM). Then, it took 4 more runs (following the first run of previous commit), over the course of more than a week, to complete training of the model. All logs, evaluation results and trained model are hereby committed.

Note:
a) Having been larger than 100 MB, two specific log files had to be removed from this commit: awesome3.err and awesome4.err, inside the results/ directory. This was allowed because the files didn't hold critical documentation data - as no special error has occurred, and the training phases in those two runs went successfully - until the jobs' times of 3 days went out (at which point te job is stopped immediately, leaving behind the last saved checkpoint to continue from).
The corresponding awesome3.out and awesome4.out, which are included, documents which epochs were completed in each run and what were the checkpoints that were saved in them, giving a full picture of what had happened in the runs despite of file restrictions.
b) A single file though was indeed critcial, and was committed in spite of being ~102 MB large. This is results/conformer_small/7775/save/CKPT+2025-04-16+17-20-59+00/optimizer.ckpt, which is part of the saved, fully-trained Conformer model; it is committed using GitHub LFS.